--- a/Cargo.toml
+++ b/Cargo.toml
@@ -39,6 +39,6 @@ crossterm = "0.28.1"
 jargon-args = "0.2.7"
 kaolinite = { path = "./kaolinite" }
 mlua = { version = "0.9.9", features = ["lua54", "vendored"] }
-quick-error = "2.0.1"
+error_set = "0.6"
 shellexpand = "3.1.0"
 synoptic = "2"