--- a/kaolinite/Cargo.toml
+++ b/kaolinite/Cargo.toml
@@ -12,7 +12,7 @@ keywords = ["unicode", "text-processing"]
 categories = ["text-processing"]
 
 [dependencies]
-quick-error = "2.0.1"
+error_set = "0.6"
 regex = "1.10.6"
 ropey = "1.6.1"
 unicode-width = "0.2"