I was looking at your code and saw that you used `quick_error`, which I hadn't heard of. I've been working on another error definition approach [error_set](https://github.com/mcmah309/error_set) and I wanted to see if there were any features in `quick_error` it was missing or I could get inspiration from. So I replaced `quick_error` with `error_set` to see. Looks like it covers all the features.

Feel free to not merge and close this PR if you prefer the ergonomics of `quick_error`.