@mcmah309 2024-10-22
Something I noticed by the way, you may have intended this, but with `quick_error` it seems you can easily for get to delegate source correctly. If not intended, `source(err)` should have been added here https://github.com/curlpipe/ox/blob/bb2192a8d526c32de784836ec1e3affb38d0bf92/src/error.rs#L12 like it is here https://github.com/curlpipe/ox/blob/bb2192a8d526c32de784836ec1e3affb38d0bf92/kaolinite/src/event.rs#L90
  - `quick_error`: `OxError::Render(ref err) => None,`
  - `error_set`: `OxError::Render(ref source) => source.source(),` (handled correctly)