@curlpipe 2024-10-22
Thank you for the PR, I really like the syntax of your `error_set` library, it feels much more compact and readable than `quick_error`. The only concern I have is that it increases the build time and number of dependencies. I've never had much of a problem with `quick_error`, so I'm hesitant to make the switch. I would love to use `error_set` if there was some way it could be made lighter.

As for the `source(err)`, that's a good catch. The code I wrote here is quite old (2-4 years old) so I can't fully remember what I intended. Either way, it's probably best to include `source` here. Thanks for bringing that to my attention.