@mcmah309 2024-10-23
Glad to hear you like it!

`error_set` is a proc macro and has no other dependencies (unless you want to enable a feature flag like `tracing`). Proc macros are all compiled beforehand, then the executable is executed by the rust compiler at build time, and since it is replacing `quick_error`, the number of dependencies will not change.

I'd be surprised if it increased the build time. Are you seeing that locally?

Either or, totally fine if you prefer the previous way. Always open to some suggestions on the API as well.