@curlpipe 2024-10-23
Yes, I tested it locally. I'll keep an eye on `error_set` and keep this PR open because I am totally open to using it. :+1: