@mcmah309 2024-10-24
I had some time today to look into the build time increase issue you mentioned and was unable to reproduce it. Is this what you are seeing as well?

### Clean build
Command:
```bash
hyperfine --prepare "cargo clean" "cargo build"
```
#### Before Changes
```console
Benchmark 1: cargo build
  Time (mean ± σ):     16.851 s ±  0.251 s    [User: 43.940 s, System: 8.691 s]
  Range (min … max):   16.604 s … 17.512 s    10 runs
```
#### After Changes
```console
Benchmark 1: cargo build
  Time (mean ± σ):     16.857 s ±  0.163 s    [User: 49.649 s, System: 9.791 s]
  Range (min … max):   16.628 s … 17.218 s    10 runs
```

### Incremental build
Command:
```bash
hyperfine --prepare 'echo -e "\nfn T$(date +"%s")(){}" >> src/main.rs' "cargo build"
```

#### Before Changes
```console
Benchmark 1: cargo build
  Time (mean ± σ):     888.3 ms ±  22.9 ms    [User: 832.3 ms, System: 713.2 ms]
  Range (min … max):   863.5 ms … 931.5 ms    10 runs
```

#### After Changes
```console
Benchmark 1: cargo build
  Time (mean ± σ):     903.8 ms ±  15.8 ms    [User: 827.6 ms, System: 739.6 ms]
  Range (min … max):   880.5 ms … 931.5 ms    10 runs
```