@curlpipe 2024-10-25
Ah yeah, you're right. Perhaps I just ended up being unlucky in my tests (I didn't test them more than once, so it might have just been a coincidence).