@curlpipe 2024-10-28
Alright, I think it's good to go.

After merge, it'll sit in the 0.6.10 branch until about next week when the 0.6.10 update is completed and published :+1: